[checkpoint_engine] use int32 for diff/gather intermediate positions
Positions are already int32 on the wire (pos_width=4); carrying them as
int64 through the rank-0 gathers doubled the position bytes for no range
benefit. Cast to int32 right after the shard-local diff (translate path)
and at converter-group assembly (NaN path), and let the gather primitives
follow the input dtype instead of hardcoding int64. Positions dominate
the sparse gather payload (8 of 10 bytes per bf16 delta), so this cuts
that traffic roughly 40%. The per-parameter <2^31 assert in
_assemble_flush already fails loud on overflow; the NaN-rebuild use site
casts back to int64 for indexing (torch index ops need Long).

+++ verl/checkpoint_engine/delta_checkpoint_engine.py
@@ -494,7 +494,7 @@ class DeltaShardedCheckpointEngine(NCCLCheckpointEngine):
                     shard_list = []
                     for gi, gv in per_rank:
                         buf = torch.full(shard_shape, float("nan"), dtype=local_dtype, device=dev)
-                        buf.view(-1)[gi] = gv
+                        buf.view(-1)[gi.to(torch.int64)] = gv  # index ops need Long
                         shard_list.append(buf)
                     for hf_name, hf_tensor in spec.to_hf(shard_list):
                         fl = hf_tensor.reshape(-1)
@@ -534,13 +534,17 @@ class DeltaShardedCheckpointEngine(NCCLCheckpointEngine):
                 # NaN-sentinel rebuild through the trainer's converter on rank 0.
                 if nan_group and nan_group[0][2] is not pg:
                     _flush_nan_group()
-                nan_group.append((name, local.dtype, pg, spec, tuple(local.shape), lidx, lval))
+                # shard-local coordinates are bounded by the shard's numel, so the
+                # full-tensor <2^31 assert in _assemble_flush covers them too.
+                nan_group.append((name, local.dtype, pg, spec, tuple(local.shape), lidx.to(torch.int32), lval))
                 if len(nan_group) >= max(batch_k, 1):
                     _flush_nan_group()
                 continue
 
             # placement profile: translate rank-locally, gather without boundaries.
-            gidx = (lidx + offset) if lidx.numel() else lidx
+            # int32 halves the gather's position bytes; the wire is int32 anyway and
+            # the per-parameter <2^31 assert in _assemble_flush covers the range.
+            gidx = ((lidx + offset) if lidx.numel() else lidx).to(torch.int32)
             gval = lval
             full_numel = _prodshape(spec.full_shape)
             if batch_k > 1:
